% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{date_tw}
\alias{date_tw}
\title{western date to tw date}
\usage{
date_tw(d_we, sep = "/")
}
\arguments{
\item{sep}{As string separate records. default is /}
}
\value{
the tw date string
}
\description{
western date to tw date
}
\examples{
  date_tw("2018-08-01")
  date_tw(as.Date("20180801", format="\%Y\%m\%d"))
}
