% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{date_we}
\alias{date_we}
\title{tw date to western date}
\usage{
date_we(d_tw, sep = "/")
}
\arguments{
\item{sep}{As string separate records. if use empty string the month and date requires
leading zero. default is /}

\item{date_tw}{An String}
}
\value{
the wester date
}
\description{
tw date to western date
}
\examples{
  date_we("1070101", sep="")
  date_we("105/02/29", sep="/")
}
