% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twdate.R
\docType{package}
\name{twdate}
\alias{twdate}
\alias{twdate-package}
\title{\pkg{twdate} an R package to translate date notations between western and taiwan}
\description{
\code{twdate} an R package to translate date notations between western and taiwan
}
