% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{year_tw}
\alias{year_tw}
\title{western year to tw year}
\usage{
year_tw(y_we)
}
\arguments{
\item{y_we}{An integer}
}
\value{
the tw year
}
\description{
western year to tw year
}
\examples{
 year_tw("2018")
 year_tw(1912)
}
