% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{year_we}
\alias{year_we}
\title{tw year to western year}
\usage{
year_we(y_tw)
}
\arguments{
\item{y_tw}{An integer}
}
\value{
the western year
}
\description{
tw year to western year
}
\examples{
  year_we(0)
  year_we("107")
}
